requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
autobahn==20.7.1
Automat==24.8.1
bcrypt==3.1.7
bidict==0.23.1
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==2.8
daphne==3.0.2
distlib==0.3.8
distro==1.7.0
dj-database-url==2.3.0
Django==4.2.14
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-user-agents==0.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
Fabric3==1.14.post1
filelock==3.13.4
Flask==2.1.0
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.5
lazr.uri==1.0.7
MarkupSafe==2.1.5
more-itertools==8.10.0
msgpack==1.1.0
multidict==6.0.5
mypy==1.11.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
netifaces==0.11.0
numpy==1.21.2
oauthlib==3.2.0
packaging==24.1
paramiko==2.12.0
pbr==6.0.0
pep8==1.7.1
pillow==10.4.0
platformdirs==4.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
SecretStorage==3.3.1
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
spark-parser==1.8.9
SQLAlchemy==2.0.29
sqlparse==0.5.0
stevedore==5.2.0
systemd-python==235
tomli==2.0.1
Twisted==22.4.0
txaio==23.1.1
typing_extensions==4.11.0
ua-parser==0.18.0
uncompyle2==2.0.0
uncompyle6==3.9.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wadllib==1.3.8
Werkzeug==2.1.0
wsproto==1.2.0
xdis==6.0.5
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1
zope.interface==7.0.3
